Added same filter support as on the comparison view to the timeline view
