+++ src/views/timeline.html
@@ -7,14 +7,35 @@
   <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery.appear/0.4.1/jquery.appear.min.js" integrity="sha256-19/eyBKQKb8IPrt7321hbLkI1+xyM8d/DpzjyiEKnCE=" crossorigin="anonymous"></script>
   <script src="/static/timeline.js" type="module"></script>
 </head>
-<body class="timeline-multi">
+<body class="timeline-multi timeline">
 
+<header>
   <div class="jumbotron">
     <h1 class="display-4">ReBench: Timeline {{project.name}}</h1>
     {{#project.description}}
     <h2 class="display-5">{{project.description}}</h2>
     {{/project.description}}
   </div>
+  <div class="refresh">
+    <div class="flex-nowrap navbar-light">
+      <button type="button" class="btn btn-outline-secondary btn-sm"
+          data-toggle="collapse" data-target="#filters" aria-controls="#filters"
+          aria-expanded="false" aria-label="Toggle Filters">
+        <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-funnel" viewBox="0 0 16 16">
+          <path d="M1.5 1.5A.5.5 0 0 1 2 1h12a.5.5 0 0 1 .5.5v2a.5.5 0 0 1-.128.334L10 8.692V13.5a.5.5 0 0 1-.342.474l-3 1A.5.5 0 0 1 6 14.5V8.692L1.628 3.834A.5.5 0 0 1 1.5 3.5v-2zm1 .5v1.308l4.372 4.858A.5.5 0 0 1 7 8.5v5.306l2-.666V8.5a.5.5 0 0 1 .128-.334L13.5 3.308V2h-11z"/>
+        </svg>
+      </button>
+    </div>
+  </div>
+</header>
+
+<div class="collapse card-columns" id="filters">
+  <div>
+    <button type="button" class="btn btn-secondary" id="filter-all">All</button>
+    <button type="button" class="btn btn-secondary" id="filter-none">None</button>
+  </div>
+  <div id="filter-groups"></div>
+</div>
 
   {{#benchmarks}}
   {{#exec}}
@@ -23,10 +44,12 @@
     <div class="title-executor">Executor: {{execName}}</div>
 
     {{#benchmarks}}
+    <div class="benchmark">
     <h4>{{benchName}}</h4>
     <p>{{cmdline}}</p>
 
     <div class="timeline-plot" data-runid="{{runId}}"></div>
+    </div>
     {{/benchmarks}}
   </div>
